CogVM source as per VMMaker.oscog-eem.1336.
Fix the hack introduced in VMMaker.oscog-eem.1199 for Sista which merely stopped
reclaiming closed PICs.  In non-Sista VMs do the usual thing of decaying usage
counts on PICs, as with methods, and reclaiming those least used.  In Sista,
retain PICs until the next cycle, identifying unused PICs as part of the
compaction scan, reminiscent of tri-colour incremental GC.

Move defaultCogCodeSize into the CogAbstractInstruction hierarchy so that
e.g.  CogARMCompiler can specify a larger default code zone.

Merge with VMMaker.oscog-tpr.1335, keeping the -1's
in the cache flush extent calculations.

Enable the magic MULL instruction on ARM.  Clean up flushICache usage

Spur:
Since the invariant is that the receiver is never forwarded, Spur must check for
forwarding on block activation.  A become between closure creation and closure
evaluation can forward the receiver, and it must be unforwarded.  Here we do so
by checking for a forwarded receiver in the value[:] primitives. This is a
major blow to block performance; e.g. the nfib block
	b := [:n| n <= 1
				ifTrue: [1]
				ifFalse: [1 + (b value: n - 1) + (b value: n - 2)]].
slows down by 11%.  So we can and will do better.  We should scan for receiver
usage in the JIT and optionally compile unforwarding code in the prolog
depending on whether self is used or not.

Cogit:
Use the Tst-based generators for immediate and SmallInteger tests more widely
to save a few more instructions.