CogVM source as per VMMaker.oscog-eem.1365
Cogit:
Fix regression in relocateMethodsPreCompaction in VMMaker.oscog-eem.1342.

Fix multiple annotations on the same pc now that annotations are directly
attached to instructions by annotating a Label with the second annotation.
Affects e.g. calls through ceNonLocalReturnTrampoline.  Neaten
generateMapAt:start: by moving the byte write inside the new
addToMap:instruction:byte:at:for:.

Fix addressIsInCurrentCompilation: using youngReferrers as the upper limit.
Revise initializeCodeZoneFrom:upTo: as a result.

ARM Cogit:
Fix relocating jumps in PICs by indirecting through the
literalsManager to skip out-of-line literals as required.

Set an out-of-line literal's opcodeIndex when dumping literals and revise the
ARM's outOfLineLiteralOpcodeLimit up a bit.  Add an assert to check that the
pc offset is in range.

Remember to nil the address on literal allocation.

Make Literals not pc-dependent; only the instructions that reference them are.