CogVM source as per VMMaker.oscog-eem.1380
Newspeak:
Enforce Newspeak access control in the interpreter, and enable it in Stack VMs.
Add a separate global lookup cache for non-ordinary sends.
Do proper lookups for self and super send misses from cogged code.
Do proper lookups for implicit receiver and outer send misses from cogged code.

Spur:
Add explicit read barriers to primitives which access an argument as the
receiver (i.e. the mirror primitives).  Don't check if the actual receiver is
used.  Simplify failure where appropriate because primitives will be retried.
Fix bogus assert in extSendBytecode.  Simplify
SpurMemoryManager>>classForClassTag:; its assert is superfluous.

General:
Eliminate a double method cache flush on startrup.

Sista Cogit:
Fix the #== bug that was crashing the VM when the counter trips in some methods.
Interesting register hack to avoid 1 move instruction in #== with forwarders.
Slighlty change the code generation of #== in sista Jit to reduce the estimated
size of abstract opcodes. The sista Jit uses 11 in instead of 10 in the regular
jit due to the counter logic, but much better than the previous value of 14.

ARM Cogit:
Fix PIC parsing for out-of-line literals.  Add an assert to the closed PIC
prototype code to check all PIC parameters are accessible.  Add a
containsAddress: to abstract away the test for a target within the PIC.