CogVM source as per VMMaker.oscog-eem.1408
ThreadedFFIPlugin:
Allow ffiCreateIntegralResultOop:ofAtomicType:in: to be inlined.  Reduce the
number of tests leading to the common case of an integral return in
ffiCalloutTo:SpecOnStack:in: etc.

Refactor the remapOop:in: idiom up into InterpreterPlugin and use it in the
ThreadedFFIPlugin.

Use an official version of MiscPrimitivePlugin.c in Newspeak (no change).