CogVM source as per VMMaker.oscog-eem.1436
Fix the bug introduced by the fix to primitive function invocation in
VMMaker.oscog-eem.1351  The fix correctly changd primitjve code to set the
primitiveFunctionPointer appropriately when a jitted external primitive was
rebound, but it didn't remember to void the jit's record of the offset of the
assignment that sets the primitiveFunctionPointer when switching between
profiling andf non-profiling regimes, so that the address from the wrong regime
would remain and be used to smash prmitive code.  The fix is simply to void the
externalSetPrimOffsets in voidCogCompiledCode.

This fixes the bug whose symptom is a hard VM crash when using
AndreasSystemProfilier.