CogVm source as per VMMaker.oscog-eem.1630
Cogit:
Fix the bug in ceCaptureStackPointers caused by pushing VarBasReg to save it
offsetting the stack pointer.

Eliminate all the maybe*VarBase methods since we have hasVarBaseRegister and
the few clients can test that and hnce not have to divine if maybeSaveVarBase
pushed something or not.

On Mac IS X modify the stack alignment definitions to reflect documented
Apple policy.