CogVM source as per VMMaker.oscog-eem.1656
Slang: Remember to remove the declaration for a local that is introduced for
inlining where the inlining uses the parse tree, not the local.  Otherwise,
stale declarations can be left that will mistype subsequent inlines of a method
with formals of the same name.  This fixes quite a few assignment type warnings.

Changes to interpreter bytecode dispatch (fetching next bytecode later)  to support optional IMMUATBILITY

primitiveStringReplace needs to check for immutability amd soon should support
16- and 64-bit access.  Provide an error code for immediate replacement.
primitiveStringReplace.

Add immutability check to changeClassOf:to: for primitiveAdoptInstance et al.

Cogit: Merge in ARM 16 bit write handling.