CogVM source as per VMMaker.oscog-eem.1664
Slang: The previous fix surfaced a bug with addTypesFor:to:in: where it would
misstate that a recursive method in the process of being typed could not be
typed because the recursive call was as yet undetermined.  The fix is simply to
discount untyped recursive calls, since we're determining precisely that type
and shouldn't abort.

And the type inference algorithm can't cope with mutual recursion, so break the
log jam by explicitly typing findMethodWithPrimitive:FromContext:UpToContext:
which is mutually recursive with findMethodWithPrimitive:FromFP:UpToContext:.