CogVM source as per VMMaker.oscog-eem.1689
ThreadedFFIPlugin:
Arrange that the plugin includes sqFFI.h, but defines ThreadedFFIPlugin as 1
before hand so as to prevent sqFFI.h from declaring the old API.  This allows
the preambleCCode to eliminate its declarations of the surface functions, which
(of course) fall foul of using ints where longts must be used on 64-bits.
Since sqFFI.h declares these functions and is included everywhere relevant
it is the gospel here.

StackInterpreter: Eliminate warnings from followForwardingPointersInStackZone:
by following the approach taken by the CoInterpreter's version.