CogVM source as per VMMaker.oscog-eem.1697
Fix serious bug in jitted perform: primitive.  Thanks to Stephane Ducasse for
finding and to Levente Uzoni for diagnosing the bug.  If the argument count
is one greater than the max number of register arguments (2 in Spur, hence
perform:with:with:) then the code to remove the receiver and arguments from
the stack neglects to remove the receiver and selector.