CogVM source as per VMMaker.oscog-eem.1699
First, it was Stephane Rollandin who found the perform: bug.  Apologies for the
misattribution.

Allow a plugin to override moduleName in generating its export table, via
moduleExportsName.  Hence fix compilation of the SqueakFFIPrims plugin as an
internal plugin.

Also use a single variable to hold moduleExportsName in the generated source.

Make the SqueakFFIPrims plugin internal in the Mac pharo builds.
Add a target to the Mac makefiles to allow printing a plugins prerequisites.
This to try and debug why plugins are not remade unless FooPlugin.lib and/or
FooPlugin.bundle are deleted.